requirements.txt
Flask>=3.0.3
supabase>=2.11.0
python-dotenv>=1.0.1
pydantic>=2.0.0
httpx>=0.24.0
postgrest>=0.10.0
asgiref>=3.7.2